Changed: Logger to better handle arguments
This was a pretty massive change.  It's not even really a refactoring.  We
completely rewrote it.

Basically the gist is that we want to be able to log not only the basic
information about the job, but also the arguments sent to the job.  This is
extremely important specifically for item processing where want to be able to
say "Hey this job took 3x as long as it needed to.  Why was that?  Let's look at
that item and see."

The `Logger` class is only the first part though.  We also need to create custom
logging middleware to replace the one that comes with Sidekiq.

--------------------------------------------------------------------------------
Change-Id:     I7cdfe50b8559999c9803a13c387405fab45f53c3
Signed-off-by: Jeff Felchner <accounts+git@thekompanee.com>