Added: Logging middleware for sidekiq
This is the thing that knows how to send the logging information to our `Logger`
class.

This class is important not only because it knows that our `Logger` prefers
things to be sent through as a hash, but also because the middleware is the only
thing that has access to the arguments that were passed to the job.

--------------------------------------------------------------------------------
Signed-off-by: Jeff Felchner <accounts+git@thekompanee.com>