Changed: Failed jobs to return the message and the backtrace
Previously, we were not doing much for failed jobs and were not returning the
message that came through. We also could not handle backtraces for failed jobs.

For backtraces, we take the first 10 lines and put them in an array which is
returned as the message.

It is also nice, because now our failed jobs have a message containing the
`error_class` and the `error_message` (ex. RuntimeError: Make this shit blow
up).